Rename Componets folder to Components
This is a simple rename, wanted to separate it from my main changes
