+++ OCR/frontend/src/components/Divider.tsx
