+++ OCR/frontend/src/components/ExtractDataHeader.tsx
