+++ OCR/frontend/src/components/ExtractStepper.tsx
