+++ OCR/frontend/src/components/FileInput.tsx
