+++ OCR/frontend/src/components/FilePreview.test.tsx
