+++ OCR/frontend/src/components/FilePreview.tsx
