+++ OCR/frontend/src/components/Header.tsx
