+++ OCR/frontend/src/components/ImageAnnotator.tsx
