+++ OCR/frontend/src/components/Stepper.tsx
