+++ OCR/frontend/src/components/UploadFile.tsx
