+++ OCR/frontend/src/components/tests/Stepper.test.tsx
