+++ OCR/frontend/src/pages/AnnotateTemplate.scss
@@ -1,4 +1,4 @@
-@import '../../node_modules/@uswds/uswds/src/stylesheets/uswds.scss';
+@import 'src/style/index.scss';
 
 .label-container {
     &:hover {
